+++ src/acceptance-test/java/org/zalando/nakadi/webservice/ConsumerLimitingServiceAT.java
@@ -5,7 +5,7 @@ import org.apache.curator.framework.CuratorFramework;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
-import org.zalando.nakadi.exceptions.NoConnectionSlotsException;
+import org.zalando.nakadi.exceptions.runtime.NoConnectionSlotsException;
 import org.zalando.nakadi.repository.zookeeper.ZooKeeperHolder;
 import org.zalando.nakadi.repository.zookeeper.ZooKeeperLockFactory;
 import org.zalando.nakadi.service.ConnectionSlot;
