+++ src/main/java/org/zalando/nakadi/controller/EventStreamController.java
@@ -29,9 +29,9 @@ import org.zalando.nakadi.domain.Timeline;
 import org.zalando.nakadi.exceptions.InternalNakadiException;
 import org.zalando.nakadi.exceptions.InvalidCursorException;
 import org.zalando.nakadi.exceptions.NakadiException;
-import org.zalando.nakadi.exceptions.NoConnectionSlotsException;
 import org.zalando.nakadi.exceptions.NoSuchEventTypeException;
 import org.zalando.nakadi.exceptions.runtime.AccessDeniedException;
+import org.zalando.nakadi.exceptions.runtime.NoConnectionSlotsException;
 import org.zalando.nakadi.exceptions.runtime.ServiceTemporarilyUnavailableException;
 import org.zalando.nakadi.exceptions.runtime.UnparseableCursorException;
 import org.zalando.nakadi.metrics.MetricUtils;
@@ -53,6 +53,7 @@ import org.zalando.nakadi.service.FeatureToggleService;
 import org.zalando.nakadi.service.timeline.TimelineService;
 import org.zalando.nakadi.util.FlowIdUtils;
 import org.zalando.nakadi.view.Cursor;
+import org.zalando.problem.MoreStatus;
 import org.zalando.problem.Problem;
 
 import javax.annotation.Nullable;
@@ -287,7 +288,8 @@ public class EventStreamController {
                 writeProblemResponse(response, outputStream, NOT_FOUND, "topic not found");
             } catch (final NoConnectionSlotsException e) {
                 LOG.debug("Connection creation failed due to exceeding max connection count");
-                writeProblemResponse(response, outputStream, e.asProblem());
+                writeProblemResponse(response, outputStream,
+                        Problem.valueOf(MoreStatus.TOO_MANY_REQUESTS, e.getMessage()));
             } catch (final ServiceTemporarilyUnavailableException e) {
                 LOG.error("Error while trying to stream events.", e);
                 writeProblemResponse(response, outputStream, SERVICE_UNAVAILABLE, e.getMessage());
