+++ src/main/java/org/zalando/nakadi/exceptions/runtime/NoConnectionSlotsException.java
@@ -1,0 +1,12 @@
+package org.zalando.nakadi.exceptions.runtime;
+
+public class NoConnectionSlotsException extends MyNakadiRuntimeException1 {
+
+    public NoConnectionSlotsException(final String msg, final Exception cause) {
+        super(msg, cause);
+    }
+
+    public NoConnectionSlotsException(final String msg) {
+        super(msg);
+    }
+}
